feat(spn): automatically set gas value and ask for approval (#488)
gas calculated depending on the simulated transaction gas cost.

* Adjust gas from the tx

* Calculate gas for the transaction

* Add confirmation prompt

* Changelog

* Not modifying factory

* Refactor spn package architecture

* Refacto broadcasting functions

* Implement prompt

* Package update

* Factorize handle broadcast error
